requirements.txt
# Flask and web
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# OpenAI
openai>=1.12.0

# LangChain and LangGraph
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
langgraph>=0.2.50

# PDF Generation
reportlab==4.1.0
Pillow==10.2.0

# Charts and Visualization
matplotlib==3.8.2
numpy==1.26.3

# Database
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
Flask-SQLAlchemy>=3.1.0

# Utilities
python-dotenv==1.0.0
pydantic>=2.5.3

# Testing
pytest==7.4.4
pytest-cov==4.1.0
